reorganise internal package
